test: move cdi tests to dedicated file
